Moved MageAttributeForm to AttributeForm in human.py
